+++ src/builder.rs
@@ -1,9 +1,10 @@
 use std::borrow::Cow;
 use std::marker::PhantomData;
 
+#[cfg(feature = "prometheus")]
 use metrics_exporter_prometheus::PrometheusHandle;
 
-use crate::{PrometheusMetricLayer, Traffic};
+use crate::{set_prefix, GenericMetricLayer, MakeDefaultHandle, Traffic};
 
 #[doc(hidden)]
 mod sealed {
@@ -20,7 +21,7 @@ impl MetricBuilderState for Paired {}
 impl MetricBuilderState for LayerOnly {}
 
 #[derive(Default, Clone)]
-/// Determines how endpoints are reported to Prometheus.
+/// Determines how endpoints are reported.
 pub enum EndpointLabel {
     /// The reported endpoint label is always the fully qualified uri path that has been requested.
     Exact,
@@ -36,8 +37,10 @@ pub enum EndpointLabel {
     MatchedPathWithFallbackFn(for<'f> fn(&'f str) -> String),
 }
 
-/// A builder for [`PrometheusMetricLayer`] that enables further customizations,
-/// such as ignoring or masking routes and defining customized [`PrometheusHandle`]s.
+/// A builder for [`GenericMetricLayer`] that enables further customizations.
+///
+/// Most of the example code uses [`PrometheusMetricLayerBuilder`], which is only a type alias
+/// specialized for Prometheus.
 ///
 /// ## Example
 /// ```rust,no_run
@@ -51,18 +54,19 @@ pub enum EndpointLabel {
 ///     .build_pair();
 /// ```
 #[derive(Clone, Default)]
-pub struct PrometheusMetricLayerBuilder<'a, S: MetricBuilderState> {
+pub struct MetricLayerBuilder<'a, T, M, S: MetricBuilderState> {
     pub(crate) traffic: Traffic<'a>,
-    pub(crate) metric_handle: Option<PrometheusHandle>,
+    pub(crate) metric_handle: Option<T>,
     pub(crate) metric_prefix: Option<String>,
-    pub(crate) _marker: PhantomData<S>,
+    pub(crate) _marker: PhantomData<(S, M)>,
 }
 
-impl<'a, S> PrometheusMetricLayerBuilder<'a, S>
+impl<'a, T, M, S> MetricLayerBuilder<'a, T, M, S>
 where
     S: MetricBuilderState,
+    M: MakeDefaultHandle,
 {
-    /// Skip reporting a specific route pattern to Prometheus.
+    /// Skip reporting a specific route pattern.
     ///
     /// In the following example
     /// ```rust
@@ -73,7 +77,7 @@ where
     ///     .build();
     /// ```
     /// any request that's URI path matches "/metrics" will be skipped altogether
-    /// when reporting to Prometheus.
+    /// when reporting to the external provider.
     ///
     /// Supports the same features as `axum`'s Router.
     ///
@@ -84,7 +88,7 @@ where
         self
     }
 
-    /// Skip reporting a collection of route patterns to Prometheus.
+    /// Skip reporting a collection of route patterns.
     ///
     /// Equivalent with calling [`with_ignore_pattern`] repeatedly.
     ///
@@ -101,7 +105,7 @@ where
     ///  _Note that ignore patterns always checked before any other group pattern rule is applied
     /// and it short-circuits if a certain route is ignored._
     ///
-    /// [`with_ignore_pattern`]: crate::PrometheusMetricLayerBuilder::with_ignore_pattern
+    /// [`with_ignore_pattern`]: crate::MetricLayerBuilder::with_ignore_pattern
     pub fn with_ignore_patterns(mut self, ignore_patterns: &'a [&'a str]) -> Self {
         self.traffic.with_ignore_patterns(ignore_patterns);
         self
@@ -133,7 +137,7 @@ where
         self
     }
 
-    /// Determine how endpoints are reported to Prometheus. For more information, see [`EndpointLabel`].
+    /// Determine how endpoints are reported. For more information, see [`EndpointLabel`].
     ///
     /// [`EndpointLabel`]: crate::EndpointLabel
     pub fn with_endpoint_label_type(mut self, endpoint_label: EndpointLabel) -> Self {
@@ -142,10 +146,13 @@ where
     }
 }
 
-impl<'a> PrometheusMetricLayerBuilder<'a, LayerOnly> {
+impl<'a, T, M> MetricLayerBuilder<'a, T, M, LayerOnly>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
     /// Initialize the builder.
-    pub fn new() -> PrometheusMetricLayerBuilder<'a, LayerOnly> {
-        PrometheusMetricLayerBuilder {
+    pub fn new() -> MetricLayerBuilder<'a, T, M, LayerOnly> {
+        MetricLayerBuilder {
             _marker: PhantomData,
             traffic: Traffic::new(),
             metric_handle: None,
@@ -153,26 +160,56 @@ impl<'a> PrometheusMetricLayerBuilder<'a, LayerOnly> {
         }
     }
 
-    /// Attach the default [`PrometheusHandle`] to the builder. This is similar to
-    /// initializing with [`PrometheusMetricLayer::pair`].
+    /// Use a prefix for the metrics instead of `axum`. This will use the following
+    /// metric names:
+    ///  - `{prefix}_http_requests_total`
+    ///  - `{prefix}_http_requests_pending`
+    ///  - `{prefix}_http_requests_duration_seconds`
+    ///
+    /// Note that this will take precedence over environment variables.
+    ///
+    /// ## Note
+    ///
+    /// This function inherently changes the metric names, beware to use the appropriate names.
+    /// There're functions in the `utils` module to get them at runtime.
+    ///
+    /// [`utils`]: crate::utils
+    pub fn with_prefix(mut self, prefix: impl Into<Cow<'a, str>>) -> Self {
+        self.metric_prefix = Some(prefix.into().into_owned());
+        self
+    }
+
+    /// Finalize the builder and get the previously registered metric handle out of it.
+    pub fn build(self) -> GenericMetricLayer<'a, T, M> {
+        GenericMetricLayer::from_builder(self)
+    }
+}
+
+impl<'a, T, M> MetricLayerBuilder<'a, T, M, LayerOnly>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
+    /// Attach the default exporter handle to the builder. This is similar to
+    /// initializing with [`GenericMetricLayer::pair`].
     ///
     /// After calling this function you can finalize with the [`build_pair`] method, and
     /// can no longer call [`build`].
     ///
-    /// [`build`]: crate::PrometheusMetricLayerBuilder::build
-    /// [`build_pair`]: crate::PrometheusMetricLayerBuilder::build_pair
-    pub fn with_default_metrics(mut self) -> PrometheusMetricLayerBuilder<'a, Paired> {
-        self.metric_handle = Some(PrometheusMetricLayer::make_default_handle());
-        PrometheusMetricLayerBuilder::<'_, Paired>::from_layer_only(self)
+    /// [`build`]: crate::MetricLayerBuilder::build
+    /// [`build_pair`]: crate::MetricLayerBuilder::build_pair
+    pub fn with_default_metrics(self) -> MetricLayerBuilder<'a, T, M, Paired> {
+        let mut builder = MetricLayerBuilder::<'_, _, _, Paired>::from_layer_only(self);
+        builder.metric_handle = Some(M::make_default_handle());
+        builder
     }
 
-    /// Attach a custom [`PrometheusHandle`] to the builder that's returned from the passed
+    /// Attach a custom built exporter handle to the builder that's returned from the passed
     /// in closure.
     ///
     /// ## Example
     /// ```rust,no_run
     /// use axum_prometheus::{
-    ///        PrometheusMetricLayerBuilder, AXUM_HTTP_REQUESTS_DURATION_SECONDS, SECONDS_DURATION_BUCKETS,
+    ///        PrometheusMetricLayerBuilder, AXUM_HTTP_REQUESTS_DURATION_SECONDS, utils::SECONDS_DURATION_BUCKETS,
     /// };
     /// use metrics_exporter_prometheus::{Matcher, PrometheusBuilder};
     ///
@@ -192,46 +229,47 @@ impl<'a> PrometheusMetricLayerBuilder<'a, LayerOnly> {
     /// After calling this function you can finalize with the [`build_pair`] method, and
     /// can no longer call [`build`].
     ///
-    /// [`build`]: crate::PrometheusMetricLayerBuilder::build
-    /// [`build_pair`]: crate::PrometheusMetricLayerBuilder::build_pair
+    /// [`build`]: crate::MetricLayerBuilder::build
+    /// [`build_pair`]: crate::MetricLayerBuilder::build_pair
     pub fn with_metrics_from_fn(
-        mut self,
-        f: impl FnOnce() -> PrometheusHandle,
-    ) -> PrometheusMetricLayerBuilder<'a, Paired> {
-        self.metric_handle = Some(f());
-        PrometheusMetricLayerBuilder::<'_, Paired>::from_layer_only(self)
-    }
-
-    /// Use a prefix for the metrics instead of `axum`. This will use the following
-    /// metric names:
-    ///  - `{prefix}_http_requests_total`
-    ///  - `{prefix}_http_requests_pending`
-    ///  - `{prefix}_http_requests_duration_seconds`
-    ///
-    /// Note that this will take precedence over environment variables.
-    pub fn with_prefix(mut self, prefix: impl Into<Cow<'a, str>>) -> Self {
-        self.metric_prefix = Some(prefix.into().into_owned());
-        self
-    }
-
-    /// Finalize the builder and get the [`PrometheusMetricLayer`] out of it.
-    pub fn build(self) -> PrometheusMetricLayer<'a> {
-        PrometheusMetricLayer::from_builder(self)
+        self,
+        f: impl FnOnce() -> T,
+    ) -> MetricLayerBuilder<'a, T, M, Paired> {
+        let mut builder = MetricLayerBuilder::<'_, _, _, Paired>::from_layer_only(self);
+        builder.metric_handle = Some(f());
+        builder
     }
 }
 
-impl<'a> PrometheusMetricLayerBuilder<'a, Paired> {
-    pub(crate) fn from_layer_only(layer_only: PrometheusMetricLayerBuilder<'a, LayerOnly>) -> Self {
-        PrometheusMetricLayerBuilder {
+impl<'a, T, M> MetricLayerBuilder<'a, T, M, Paired>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
+    pub(crate) fn from_layer_only(layer_only: MetricLayerBuilder<'a, T, M, LayerOnly>) -> Self {
+        if let Some(prefix) = layer_only.metric_prefix.as_ref() {
+            set_prefix(prefix);
+        }
+        MetricLayerBuilder {
             _marker: PhantomData,
             traffic: layer_only.traffic,
             metric_handle: layer_only.metric_handle,
             metric_prefix: layer_only.metric_prefix,
         }
     }
-    /// Finalize the builder and get out the [`PrometheusMetricLayer`] and the
-    /// [`PrometheusHandle`] out of it as a tuple.
-    pub fn build_pair(self) -> (PrometheusMetricLayer<'a>, PrometheusHandle) {
-        PrometheusMetricLayer::pair_from_builder(self)
+}
+
+impl<'a, T, M> MetricLayerBuilder<'a, T, M, Paired>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
+    /// Finalize the builder and get out the [`GenericMetricLayer`] and the
+    /// exporter handle out of it as a tuple.
+    pub fn build_pair(self) -> (GenericMetricLayer<'a, T, M>, T) {
+        GenericMetricLayer::pair_from_builder(self)
     }
 }
+
+#[cfg(feature = "prometheus")]
+/// A builder for [`crate::PrometheusMetricLayer`] that enables further customizations.
+pub type PrometheusMetricLayerBuilder<'a, S> =
+    MetricLayerBuilder<'a, PrometheusHandle, crate::Handle, S>;