+++ src/lib.rs
@@ -1,6 +1,6 @@
-//! A Prometheus middleware to collect HTTP metrics for Axum applications.
+//!A middleware to collect HTTP metrics for Axum applications.
 //!
-//! `axum-prometheus` relies on [`metrics_exporter_prometheus`] as a backed to interact with Prometheus.
+//! `axum-prometheus` relies on [`metrics.rs`](https://metrics.rs/) and its ecosystem to collect and export metrics - for instance for Prometheus, `metrics_exporter_prometheus` is used as a backed to interact with Prometheus.
 //!
 //! ## Metrics
 //!
@@ -87,6 +87,51 @@
 //! axum_http_requests_duration_seconds_count{method="GET",status="200",endpoint="/metrics"} 4
 //! ```
 //!
+//! ## Using a different exporter than Prometheus
+//!
+//! This crate may be used with other exporters than Prometheus. First, disable the default features:
+//!
+//! ```toml
+//! axum-prometheus = { version = "0.3.4", default-features = false }
+//! ```
+//!
+//! Then implement the `MakeDefaultHandle` for the provider you'd like to use. For `StatsD`:
+//!
+//! ```rust,ignore
+//! use metrics_exporter_statsd::StatsdBuilder;
+//! use axum_prometheus::{MakeDefaultHandle, GenericMetricLayer};
+//!
+//! // A marker struct for the custom StatsD exporter.
+//! struct Recorder;
+//!
+//! // In order to use this with `axum_prometheus`, we must implement `MakeDefaultHandle`.
+//! impl MakeDefaultHandle for Recorder {
+//!     type Out = ();
+//!
+//!     fn make_default_handle() -> Self::Out {
+//!         // The regular setup for StatsD..
+//!         let recorder = StatsdBuilder::from("127.0.0.1", 8125)
+//!             .with_queue_size(5000)
+//!             .with_buffer_size(1024)
+//!             .build(Some("prefix"))
+//!             .expect("Could not create StatsdRecorder");
+//!
+//!         metrics::set_boxed_recorder(Box::new(recorder)).unwrap();
+//!         // We don't need to return anything meaningful from here (unlike PrometheusHandle)
+//!         // Let's just return an empty tuple.
+//!         ()
+//!     }
+//! }
+//!
+//! fn main() {
+//!     // ...
+//!     // Use `GenericMetricLayer` instead of `PrometheusMetricLayer`.
+//!     let (metric_layer, _handle) = GenericMetricLayer::<'_, _, Recorder>::pair();
+//!     // ...
+//!
+//! }
+//! ```
+//!
 //! This crate is similar to (and takes inspiration from) [`actix-web-prom`](https://github.com/nlopes/actix-web-prom) and [`rocket_prometheus`](https://github.com/sd2k/rocket_prometheus),
 //! and also builds on top of davidpdrsn's [earlier work with LifeCycleHooks](https://github.com/tower-rs/tower-http/pull/96) in `tower-http`.
 //!
@@ -126,13 +171,16 @@ pub static PREFIXED_HTTP_REQUESTS_PENDING: OnceCell<String> = OnceCell::new();
 
 use std::borrow::Cow;
 use std::collections::HashMap;
+use std::marker::PhantomData;
 use std::time::Instant;
 
 mod builder;
 pub mod lifecycle;
-mod utils;
+pub mod utils;
 use axum::extract::MatchedPath;
 pub use builder::EndpointLabel;
+pub use builder::MetricLayerBuilder;
+#[cfg(feature = "prometheus")]
 pub use builder::PrometheusMetricLayerBuilder;
 use builder::{LayerOnly, Paired};
 use lifecycle::layer::LifeCycleLayer;
@@ -141,11 +189,12 @@ use metrics::{decrement_gauge, histogram, increment_counter, increment_gauge};
 use once_cell::sync::OnceCell;
 use tower::Layer;
 use tower_http::classify::{ClassifiedResponse, SharedClassifier, StatusInRangeAsFailures};
-pub use utils::SECONDS_DURATION_BUCKETS;
 
+#[cfg(feature = "prometheus")]
 use metrics_exporter_prometheus::{Matcher, PrometheusBuilder, PrometheusHandle};
 
 pub use metrics;
+#[cfg(feature = "prometheus")]
 pub use metrics_exporter_prometheus;
 
 /// Use a prefix for the metrics instead of `axum`. This will use the following
@@ -329,17 +378,31 @@ impl<'a, FailureClass> Callbacks<FailureClass> for Traffic<'a> {
     }
 }
 
-/// The tower middleware layer for recording http metrics with Prometheus.
-#[derive(Clone)]
-pub struct PrometheusMetricLayer<'a> {
+/// The tower middleware layer for recording http metrics with different exporters.
+pub struct GenericMetricLayer<'a, T, M> {
     pub(crate) inner_layer: LifeCycleLayer<SharedClassifier<StatusInRangeAsFailures>, Traffic<'a>>,
+    _marker: PhantomData<(T, M)>,
 }
 
-impl<'a> PrometheusMetricLayer<'a> {
-    /// Create a new tower middleware that can be used to track metrics with Prometheus.
+// We don't require that `T` nor `M` is `Clone`, since none of them is actually contained in this type.
+impl<'a, T, M> std::clone::Clone for GenericMetricLayer<'a, T, M> {
+    fn clone(&self) -> Self {
+        GenericMetricLayer {
+            inner_layer: self.inner_layer.clone(),
+            _marker: self._marker.clone(),
+        }
+    }
+}
+
+impl<'a, T, M> GenericMetricLayer<'a, T, M>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
+    /// Create a new tower middleware that can be used to track metrics.
     ///
     /// By default, this __will not__ "install" the exporter which sets it as the
-    /// global recorder for all `metrics` calls. Instead, here you can use the [`metrics_exporter_prometheus::PrometheusBuilder`]
+    /// global recorder for all `metrics` calls.
+    /// If you're using Prometheus, here you can use [`metrics_exporter_prometheus::PrometheusBuilder`]
     /// to build your own customized metrics exporter.
     ///
     /// This middleware is using the following constants for identifying different HTTP metrics:
@@ -354,7 +417,7 @@ impl<'a> PrometheusMetricLayer<'a> {
     /// # Example
     /// ```
     /// use axum::{routing::get, Router};
-    /// use axum_prometheus::{AXUM_HTTP_REQUESTS_DURATION_SECONDS, SECONDS_DURATION_BUCKETS, PrometheusMetricLayer};
+    /// use axum_prometheus::{AXUM_HTTP_REQUESTS_DURATION_SECONDS, utils::SECONDS_DURATION_BUCKETS, PrometheusMetricLayer};
     /// use metrics_exporter_prometheus::{Matcher, PrometheusBuilder};
     /// use std::net::SocketAddr;
     ///
@@ -393,36 +456,33 @@ impl<'a> PrometheusMetricLayer<'a> {
         let make_classifier =
             StatusInRangeAsFailures::new_for_client_and_server_errors().into_make_classifier();
         let inner_layer = LifeCycleLayer::new(make_classifier, Traffic::new());
-        Self { inner_layer }
+        Self {
+            inner_layer,
+            _marker: PhantomData,
+        }
     }
 
-    pub(crate) fn from_builder(builder: PrometheusMetricLayerBuilder<'a, LayerOnly>) -> Self {
-        if let Some(prefix) = builder.metric_prefix {
-            set_prefix(prefix);
-        }
-
+    pub(crate) fn from_builder(builder: MetricLayerBuilder<'a, T, M, LayerOnly>) -> Self {
         let make_classifier =
             StatusInRangeAsFailures::new_for_client_and_server_errors().into_make_classifier();
         let inner_layer = LifeCycleLayer::new(make_classifier, builder.traffic);
-        Self { inner_layer }
+        Self {
+            inner_layer,
+            _marker: PhantomData,
+        }
     }
 
-    pub(crate) fn pair_from_builder(
-        builder: PrometheusMetricLayerBuilder<'a, Paired>,
-    ) -> (Self, PrometheusHandle) {
-        if let Some(prefix) = builder.metric_prefix {
-            set_prefix(prefix);
-        }
-
+    pub(crate) fn pair_from_builder(builder: MetricLayerBuilder<'a, T, M, Paired>) -> (Self, T) {
         let make_classifier =
             StatusInRangeAsFailures::new_for_client_and_server_errors().into_make_classifier();
         let inner_layer = LifeCycleLayer::new(make_classifier, builder.traffic);
 
         (
-            Self { inner_layer },
-            builder
-                .metric_handle
-                .unwrap_or_else(Self::make_default_handle),
+            Self {
+                inner_layer,
+                _marker: PhantomData,
+            },
+            builder.metric_handle.unwrap_or_else(M::make_default_handle),
         )
     }
 
@@ -456,11 +516,76 @@ impl<'a> PrometheusMetricLayer<'a> {
     ///    // server.await.unwrap();
     /// }
     /// ```
-    pub fn pair() -> (Self, PrometheusHandle) {
-        (Self::new(), Self::make_default_handle())
+    pub fn pair() -> (Self, T) {
+        (Self::new(), M::make_default_handle())
     }
+}
 
-    pub(crate) fn make_default_handle() -> PrometheusHandle {
+impl<'a, T, M> Default for GenericMetricLayer<'a, T, M>
+where
+    M: MakeDefaultHandle<Out = T>,
+{
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl<'a, S, T, M> Layer<S> for GenericMetricLayer<'a, T, M> {
+    type Service = LifeCycle<S, SharedClassifier<StatusInRangeAsFailures>, Traffic<'a>>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        self.inner_layer.layer(inner)
+    }
+}
+
+/// The trait that allows to use a metrics exporter in `GenericMetricLayer`.
+pub trait MakeDefaultHandle {
+    /// The type of the metrics handle to return from [`MetricLayerBuilder`].
+    type Out;
+
+    /// The function that defines how to initialize a metric exporter by default, if none were provided.
+    ///
+    /// # Example
+    ///
+    /// ```rust, no_run
+    /// use metrics_exporter_prometheus::{PrometheusBuilder, PrometheusHandle, Matcher};
+    /// use axum_prometheus::{utils::{SECONDS_DURATION_BUCKETS, requests_duration_name}, MakeDefaultHandle, GenericMetricLayer};
+    ///
+    /// // A wrapper struct to work around Rust's orphan rules.
+    /// pub struct Handle(pub PrometheusHandle);
+    ///
+    /// impl MakeDefaultHandle for Handle {
+    ///     type Out = PrometheusHandle;
+    ///
+    ///     fn make_default_handle() -> Self::Out {
+    ///         PrometheusBuilder::new()
+    ///             .set_buckets_for_metric(
+    ///                 Matcher::Full(requests_duration_name().to_string()),
+    ///                 SECONDS_DURATION_BUCKETS,
+    ///             )
+    ///             .unwrap()
+    ///             .install_recorder()
+    ///             .unwrap()
+    ///     }
+    /// }
+    /// ```
+    /// and then, to use it:
+    /// ```rust,ignore
+    /// let (layer, handle) =  GenericMetricLayer::<'_, _, Handle>::pair();
+    /// ```
+    fn make_default_handle() -> Self::Out;
+}
+
+/// The default handle for the Prometheus exporter.
+#[cfg(feature = "prometheus")]
+#[derive(Clone)]
+pub struct Handle(pub PrometheusHandle);
+
+#[cfg(feature = "prometheus")]
+impl MakeDefaultHandle for Handle {
+    type Out = PrometheusHandle;
+
+    fn make_default_handle() -> Self::Out {
         PrometheusBuilder::new()
             .set_buckets_for_metric(
                 Matcher::Full(
@@ -469,7 +594,7 @@ impl<'a> PrometheusMetricLayer<'a> {
                         .map_or(AXUM_HTTP_REQUESTS_DURATION_SECONDS, |s| s.as_str())
                         .to_string(),
                 ),
-                SECONDS_DURATION_BUCKETS,
+                utils::SECONDS_DURATION_BUCKETS,
             )
             .unwrap()
             .install_recorder()
@@ -477,16 +602,6 @@ impl<'a> PrometheusMetricLayer<'a> {
     }
 }
 
-impl<'a> Default for PrometheusMetricLayer<'a> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<'a, S> Layer<S> for PrometheusMetricLayer<'a> {
-    type Service = LifeCycle<S, SharedClassifier<StatusInRangeAsFailures>, Traffic<'a>>;
-
-    fn layer(&self, inner: S) -> Self::Service {
-        self.inner_layer.layer(inner)
-    }
-}
+#[cfg(feature = "prometheus")]
+/// The tower middleware layer for recording http metrics with Prometheus.
+pub type PrometheusMetricLayer<'a> = GenericMetricLayer<'a, PrometheusHandle, Handle>;