+++ src/utils.rs
@@ -1,5 +1,11 @@
 use http::Method;
 
+use crate::{
+    AXUM_HTTP_REQUESTS_DURATION_SECONDS, AXUM_HTTP_REQUESTS_PENDING, AXUM_HTTP_REQUESTS_TOTAL,
+    PREFIXED_HTTP_REQUESTS_DURATION_SECONDS, PREFIXED_HTTP_REQUESTS_PENDING,
+    PREFIXED_HTTP_REQUESTS_TOTAL,
+};
+
 /// Standard HTTP request duration buckets measured in seconds. The default buckets are tailored to broadly
 /// measure the response time of a network service. Most likely, however, you will be required to define
 /// buckets customized to your use case.
@@ -20,4 +26,34 @@ pub(super) const fn as_label(method: &Method) -> &'static str {
         Method::PATCH => "PATCH",
         _ => "",
     }
+}
+
+/// The name of the requests total metric. By default, it's the same as [`AXUM_HTTP_REQUESTS_TOTAL`], but
+/// can be changed via the [`with_prefix`] function.
+///
+/// [`with_prefix`]: crate::MetricLayerBuilder::with_prefix
+pub fn requests_total_name() -> &'static str {
+    PREFIXED_HTTP_REQUESTS_TOTAL
+        .get()
+        .map_or(AXUM_HTTP_REQUESTS_TOTAL, |s| s.as_str())
+}
+
+/// The name of the requests duration metric. By default, it's the same as [`AXUM_HTTP_REQUESTS_DURATION_SECONDS`], but
+/// can be changed via the [`with_prefix`] function.
+///
+/// [`with_prefix`]: crate::MetricLayerBuilder::with_prefix
+pub fn requests_duration_name() -> &'static str {
+    PREFIXED_HTTP_REQUESTS_DURATION_SECONDS
+        .get()
+        .map_or(AXUM_HTTP_REQUESTS_DURATION_SECONDS, |s| s.as_str())
+}
+
+/// The name of the requests pending metric. By default, it's the same as [`AXUM_HTTP_REQUESTS_PENDING`], but
+/// can be changed via the [`with_prefix`] function.
+///
+/// [`with_prefix`]: crate::MetricLayerBuilder::with_prefix
+pub fn requests_pending_name() -> &'static str {
+    PREFIXED_HTTP_REQUESTS_PENDING
+        .get()
+        .map_or(AXUM_HTTP_REQUESTS_PENDING, |s| s.as_str())
 }