+++ tests/common.rs
@@ -1,0 +1,7 @@
+use http::{Request, Response};
+use hyper::Body;
+use tower::BoxError;
+
+pub async fn echo(req: Request<Body>) -> Result<Response<Body>, BoxError> {
+    Ok(Response::new(req.into_body()))
+}